Working C# version of Battleship with bugs

+++ battleships/src/Model/Ship.cs
@@ -1,4 +1,3 @@
-
 using Microsoft.VisualBasic;
 using System;
 using System.Collections;
@@ -34,7 +33,7 @@ public class Ship
 				return "Aircraft Carrier";
 			}
 
-			return _shipName.ToString();
+			return _shipName.ToString ();
 		}
 	}
 
@@ -74,36 +73,36 @@ public class Ship
 		get { return _direction; }
 	}
 
-	public Ship(ShipName ship)
+	public Ship (ShipName ship)
 	{
 		_shipName = ship;
-		_tiles = new List<Tile>();
+		_tiles = new List<Tile> ();
 
 		//gets the ship size from the enumarator
-		_sizeOfShip = _shipName;
+		_sizeOfShip = (int)_shipName;
 	}
 
 	/// <summary>
 	/// Add tile adds the ship tile
 	/// </summary>
 	/// <param name="tile">one of the tiles the ship is on</param>
-	public void AddTile(Tile tile)
+	public void AddTile (Tile tile)
 	{
-		_tiles.Add(tile);
+		_tiles.Add (tile);
 	}
 
 	/// <summary>
 	/// Remove clears the tile back to a sea tile
 	/// </summary>
-	public void Remove()
+	public void Remove ()
 	{
 		foreach (Tile tile in _tiles) {
-			tile.ClearShip();
+			tile.ClearShip ();
 		}
-		_tiles.Clear();
+		_tiles.Clear ();
 	}
 
-	public void Hit()
+	public void Hit ()
 	{
 		_hitsTaken = _hitsTaken + 1;
 	}
@@ -126,17 +125,10 @@ public class Ship
 	/// <param name="direction"></param>
 	/// <param name="row"></param>
 	/// <param name="col"></param>
-	internal void Deployed(Direction direction, int row, int col)
+	internal void Deployed (Direction direction, int row, int col)
 	{
 		_row = row;
 		_col = col;
 		_direction = direction;
 	}
 }
-
-//=======================================================
-//Service provided by Telerik (www.telerik.com)
-//Conversion powered by NRefactory.
-//Twitter: @telerik
-//Facebook: facebook.com/telerik
-//=======================================================
